Partial N fix
Tracked down the problem to the N quadrature. There is a discontinuous derivative in Ni @ x+y*T=0, which leads to noise. I have reduced the problem by shifting the discontinuity to the center, but the shift is x/y, which gets too big. I should also adjust the scale.